+++ packages/shared_preferences/example/android/app/src/main/java/dev/flutter/plugins/sharedpreferencesexample/MainActivity.java
@@ -4,9 +4,9 @@
 
 package dev.flutter.plugins.sharedpreferencesexample;
 
-import dev.flutter.plugins.sharedpreferences.SharedPreferencesPlugin;
 import io.flutter.embedding.android.FlutterActivity;
 import io.flutter.embedding.engine.FlutterEngine;
+import io.flutter.plugins.sharedpreferences.SharedPreferencesPlugin;
 
 public class MainActivity extends FlutterActivity {
